Get SLURM services running again.
This required a rather nasty hack.
By default, all services are now run under jovyan user.
Unfortunately, SLURM really wants to be run under root.

We can run the container under root, but the the
20_start-postgres.sh script fails.

Hence, we need to hack the start.sh
script to ensure that all the hooks in before-notebook.d/
are executed by jovyan. Scripts in start-notebook.d/
(currently empty) will be executed by root, so that's
where we place the slurm daemons, executed by `run-one-continuously`