Fix rabbitmq startup
I have no idea why rabbitmq-server fails to start
when run through sudo (the error that it gives is very cryptic),
but it seems that using 'runuser' instead of sudo works.

This is all still a horrible hack so that we can run slurm
services under root. I am not happy about it
but I am out of ideas how to do it more cleanly,
while still using the aiidalab/full-stack image.